% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greta-interface.R
\name{greta_interface}
\alias{greta_interface}
\alias{initialise}
\alias{initialize}
\alias{parameters}
\title{work with greta objects created with \link[greta.integrated]{integrated_model}}
\usage{
initialise(object)

initialize(object)

parameters(object, ...)
}
\arguments{
\item{object}{a \link[greta]{greta_model} object}
}
\value{
A \link[greta]{initials} object
}
\description{
A set of helper functions.
}
\details{
something
}
\examples{
\dontrun{

 library(integrated)

# prepare an example model
model <- integrated_model()
                        
# summarise fitted model
model
summary(model)
plot(model)
}
}
