% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_density-class.R
\name{integrated_density}
\alias{integrated_density}
\alias{ricker}
\alias{beverton}
\alias{no_density}
\alias{is.integrated_density}
\alias{print.integrated_density}
\alias{summary.integrated_density}
\alias{plot.integrated_density}
\title{create integrated density objects}
\usage{
ricker(parameters = uniform(0, 1))

beverton(parameters = uniform(0, 1))

no_density(parameters = NULL)

is.integrated_density(object)

\method{print}{integrated_density}(x, ...)

\method{summary}{integrated_density}(object, ...)

\method{plot}{integrated_density}(x, ...)
}
\arguments{
\item{object}{an \code{integrated_density} object}

\item{x}{an \code{integrated_density} object}

\item{...}{additional arguments to \link[base]{print}, \link[base]{summary}, and \link[graphics]{plot} methods (currently ignored)}

\item{lambda}{parameter of Ricker and Beverton-Holt models of density dependence}
}
\value{
An object of class \code{integrated_density}, which can be used to create
   and \link[greta.integrated]{integrated_process} object
}
\description{
An \code{integrated_density} object specifies the form of density
  dependence and associated prior distributions for an integrated population analysis
}
\details{
Prior distributions can be specified as single-dimensional
  greta distribution, e.g., \code{normal(0, 1)}. Link functions and transformations
  can be specified directly in-line, e.g., \code{ilogit(normal(0, 1))} specifies
  normal priors with a mean of zero and a standard deviation of one, transformed
  with an inverse-logit link.
}
\examples{
\dontrun{

library(integrated)

# a really basic age-structured model with five age classes
process <- leslie(5, density = ricker(lambda = uniform(0, 1)))

# setting custom priors
process <- leslie(5, density = beverton(lambda = uniform(0, 1)),
                  priors = list(survival = ilogit(normal(0, 1)),
                                fecundity = exp(normal(0, 1))))
}
}
