% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_predictor-class.R
\name{integrated_predictor}
\alias{integrated_predictor}
\alias{add_predictors}
\alias{is.integrated_predictor}
\alias{print.integrated_predictor}
\alias{summary.integrated_predictor}
\title{create integrated predictor objects}
\usage{
add_predictors(formula, data)

is.integrated_predictor(object)

\method{print}{integrated_predictor}(x, ...)

\method{summary}{integrated_predictor}(object, ...)
}
\arguments{
\item{formula}{formula specifying the predictor structure (see \link[lme4]{lmer} for details)}

\item{data}{a single data input (see details for descriptions of possible input types)}

\item{object}{an \code{integrated_data} object}

\item{x}{an \code{integrated_data} object}

\item{...}{additional arguments to \link[base]{print} and \link[base]{summary} methods (currently ignored)}
}
\value{
An object of class \code{integrated_predictor}, which contains information on predictors and
  can be passed to an \link[greta.integrated]{integrated_data} constructor.
}
\description{
\code{integrated_predictor} defines a data object containing predictor data
  that can be passed to an \link[greta.integrated]{integrated_data} constructor.
}
\details{
Add something.
}
\examples{
\dontrun{

library(integrated)

# prepare an example model
data <- add_data()
                        
# summarise data module
model
summary(model)
plot(model)
}
}
