% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_summary-class.R
\name{integrated_summary}
\alias{integrated_summary}
\title{summarise greta models containing \link[greta.integrated]{integrated_model} components}
\usage{
integrated_summary(object)
}
\arguments{
\item{object}{a \link[coda]{mcmc.list} object created with \link[greta]{mcmc}}
}
\value{
A \link[greta.integrated]{integrated_summary} object
}
\description{
A set of helper functions.
}
\details{
something
}
\examples{
\dontrun{

 library(integrated)

# prepare an example model
model <- integrated_model()
                        
# summarise fitted model
model
summary(model)
plot(model)
}
}
