% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_data-class.R
\name{integrated_data}
\alias{integrated_data}
\alias{age_abundance}
\alias{stage_abundance}
\alias{age_recapture}
\alias{stage_recapture}
\alias{stage_to_age}
\alias{age_to_stage}
\alias{community}
\alias{predictors}
\alias{is.integrated_data}
\alias{print.integrated_data}
\alias{summary.integrated_data}
\title{create integrated data objects}
\usage{
age_abundance(data, process, bias = no_bias(), settings = list())

stage_abundance(data, process, bias = no_bias(), settings = list())

age_recapture(data, process, bias = no_bias(), settings = list())

stage_recapture(data, process, bias = no_bias(), settings = list())

stage_to_age(x, ...)

age_to_stage(x, ...)

community(data, process, bias = no_bias(), settings = list())

predictors(data, process, bias = no_bias(), settings = list())

is.integrated_data(object)

\method{print}{integrated_data}(x, ...)

\method{summary}{integrated_data}(object, ...)
}
\arguments{
\item{data}{a single data input (see details for descriptions of possible input types)}

\item{process}{an \link[greta.integrated]{integrated_process} object}

\item{bias}{a bias function that connects observed and modelled data (see details)}

\item{settings}{a named list of settings passed to data formatting functions (see details)}

\item{x}{an \code{integrated_data} object}

\item{...}{additional arguments to \link[base]{print} and \link[base]{summary} methods (currently ignored)}

\item{object}{an \code{integrated_data} object}
}
\value{
An object of class \code{integrated_data}, which contains information on the data module and
  can be passed to \link[greta.integrated]{integrated_model}
}
\description{
\code{integrated_data} defines a data object with custom likelihood based on
 a process model defined with \link[greta.integrated]{integrated_process}
}
\details{
Do something. The settings list can be used to specify how the data are binned, either with
  specific breaks for binning or with the number of breaks to use. If these are not provided, the 
  functions use the \code{classes} element of \code{process} to determine the number
  of bins (\code{nbreaks = classes + 1}).
}
\examples{
\dontrun{

library(integrated)

# prepare an example model
data <- add_data()
                        
# summarise data module
model
summary(model)
plot(model)
}
}
