% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greta-interface.R, R/integrated_model-class.R
\name{greta_interface}
\alias{greta_interface}
\alias{initialise}
\alias{integrated_model}
\alias{is.integrated_model}
\alias{print.integrated_model}
\alias{plot.integrated_model}
\alias{summary.integrated_model}
\title{work with greta objects created with \link[greta.integrated]{integrated_model}}
\usage{
initialise(object)

integrated_model(...)

is.integrated_model(object)

\method{print}{integrated_model}(x, ...)

\method{plot}{integrated_model}(x, ...)

\method{summary}{integrated_model}(object, ...)
}
\arguments{
\item{object}{a \link[greta]{greta_model} object}

\item{...}{one or more \link[greta.integrated]{integrated_data} objects}

\item{x}{an \code{integrated_model} object}

\item{process}{an \link[greta.integrated]{integrated_process} object}

\item{object}{an \code{integrated_model} object}
}
\value{
A \link[greta]{initials} object

A named list of \code{greta_array} objects, which can be passed to
   \link[greta]{model}. This list has associated `print`, `plot`, and `summary` methods.
}
\description{
A set of helper functions.

An \code{integrated_model} object combines multiple \link[greta.integrated]{integrated_data}
  objects with one or more \link[greta.integrated]{integrated_process} objects and returns all
   \link[greta]{greta_array} objects required to fit an integrated population model. Model fitting
   is handled separately with the \link[greta]{model} and \link[greta]{mcmc} functions.
}
\details{
something

something
}
\examples{
\dontrun{

 library(integrated)

# prepare an example model
model <- integrated_model()
                        
# summarise fitted model
model
summary(model)
plot(model)
}
\dontrun{

### ADD EXAMPLES OF ALL FUNCTIONS HERE

library(integrated)

# prepare an example model
model <- integrated_model()
                        
# summarise fitted model
model
summary(model)
plot(model)
}
}
